initial tests and refinement of existing base solution

+++ src/EtherDLLClient.hpp
@@ -604,7 +604,7 @@ public:
 	{
 		const std::string logSource = "pingClient";
 
-		int pingPeriodMs = config[service::KEY][service::PingPeriod::KEY].get<double>() * 1000;
+		int pingPeriodMs = static_cast<int>(config[service::KEY][service::PingPeriod::KEY].get<double>() * 1000);
 		int iResult = 0;
 
 		while (interruptionCode == edll::Code::RUNNING) {
